This adds a module called imphot (Imaging Photometry)
The imphot module provides python functions and scripts for fitting
for the photometric parameters of MUSE images, using two methods that
work by comparing MUSE images to HST images of the same region. The
module includes extensive documentation, both as web-pages rendered by
sphinx, and as docstrings and comments within the source-code.