Rewrote the pitfalls page again to incorporate new findings on stellar flux losses.
It turns out that the flux-scaling discrepancy that is seen between
HST and MUSE images of point sources, is a known problem of the HST
ACS, caused by CTE charge loss. This effect predominantly affected
faint stars (and QSOs) against dark backgrounds, as described by Riess
in an ACS science report from 2003.