For HST images without DATA or SCI extensions, force MPDAF to read the primary data array.
If MPDAF can't find a DATA or SCI extension in a FITS file, it
complains and crashes, unless an extension is explicitly specified by
the user. It really ought to just read the primary data array in this
case, but an MPDAF patch to do this was rejected.  In
regrid_hst_to_muse, this was a problem because HST FITS files don't
have either of these extensions. A workaround has been added that
passes MPDAF ext=0 if neither of these extensions exist.