Increase the speed of make_wideband_image by using multiple processes.
The make_wideband_image program needed to be made much faster, because
it would have taken 26 hours to use it to process all of the MUSE
exposures. Specifically, it took 85 seconds for it to generate an
image from each MUSE cube (on urania4), and there are currently 279
MUSE cubes to be processed, and 4 filters for which images are needed.

To improve the speed, instead of calling bandpass_image() in
MPDAF, which only uses one process, a new function was written
that performs the equivalent operation using one process per CPU
core. On urania4, for each input cube this function takes 20
seconds to produce the first of the filtered images (dominated by
the time taken by the OS to page in the file), then 4 seconds each
to produce images for the remaining 3 filters. So in total it
should take 279*32 seconds, or 2.5 hours to generate all of the
images.

The new function is also mentioned in the documentation, in place
of the MPDAF bandpass_image() function.

The new function would be easy to port to MPDAF, since it has the
same arguments as the MPDAF function, except for taking a cube
argument instead of self. However all of the multiprocessing
classes etc, include quite a bit of code, and MPDAF already takes
4 seconds to import, so this may not be worth doing.