In imphot.fitimage, subtract the median of the muse image before fitting.
In imphot.fitimage(), masked areas are zeroed before fitting, so that
they don't affect the fitted scale factor. For MUSE images with an
average background level of zero, this works well. However for images
with strong sky backgrounds, the masked areas look like deep holes in
the background, compared to the unmasked areas. These strong artefacts
dominate the residuals, so the fitting algorithm fails. To avoid this
problem, the revised code measures the median pixel value of the MUSE
image, and subtracts this from the MUSE image before zeroing the
masked areas and performing the fit. This suppresses the large steps
between unmasked and masked regions.  The fitted background offset
includes the effect of the subtracted median value, and the subtracted
median value is used as the starting value for the fitted background
offset.