Parameterize multiprocessing in imphot make_wideband_image and bandpass_image
The imphot make_wideband_image script and the imphot.band_pass_image()
function normally fork multiple worker processes to spread the work
across multiple CPUs. An optional 'nprocess' argument has been added
that lets the user specify the maximum number of worker processes (in
addition to the main process).  As a special case, no worker processes
are used if 1 is passed to this option.