In imphot fit_image_photometry() refine fits with large fitted position offsets.
In fit_image_photometry(), if the position offset of the input images
is larger than a pixel, then the pixel mask that is applied to the two
images does not mask the same regions of the sky in the two
images. So, when the fitted position x or y offsets are more than a
pixel different than the initial guesses, repeat the fit with the
fitted position offsets as the new initial guess.