Provide a way to apply an external mask when fitting images with imphot.
fit_image_photometry() now has an extramask argument, which can be
used to specify an extra masking image as a FITS filename, an Image()
object, or a 2D numpy array. The fit_photometry script has an
equivalent optional argument called --extramask, which takes the name
of a FITS file.

This option is currently needed so that satellite trails can be masked
using custom masks that are currently stored in FITS files.